windows: the MSVC build was broken, and it was one missing include
MSVC has never been run against these sources and could not be run here, so the
audit went the other way: seven passes over the 21 translation units in the
airusb-net target, each finding then adversarially re-checked against the real
microsoft/STL sources rather than against recollection.

One genuine build break existed. crypto/Primitives.h declared
`std::string toHex(...)` while including <string_view> but not <string>. libc++
and libstdc++ both leak <string> through other headers, which is the entire
reason Clang and MinGW have been silent about it. Microsoft's does not:
<string_view> reaches <xstring> only under the opt-in
_LEGACY_CODE_ASSUMES_STRING_VIEW_INCLUDES_XSTRING, and std::string's alias lives
nowhere else. Seven of the ten TUs that reach that header had no prior <string>,
and crypto/Primitives.cpp is the first source of the first library the documented
command builds — so `cmake --build --target airusb-net` died almost immediately.
One line fixes it.

Two latent defects, neither Windows-specific, found on the way:

  - ImporterClient took subspan(40) of an ATTACH_OK body whose length nothing had
    checked. call() validates replies only far enough to know body_len matches the
    bytes present; it never runs validateHeader on a reply. A peer answering
    status != Ok with body_len == 0 wrapped that to a 2^64-sized span over a null
    pointer and forEachTlv then read it. In-tree exporters always send the 40
    bytes, so this needed a hostile or broken peer — at an address a user types.
  - The handshake timeout counted iterations of sleepMs(1) and called the count
    milliseconds. Windows' sleep tick is 15.625 ms, so the documented 15 s window
    was about four minutes there, pinning a single-threaded client on a half-open
    peer. It now reads the continuous clock, which also survives suspend.

What the audit did NOT find is the more useful half, and WINDOWS.md records it so
nobody re-derives it: SO_REUSEADDR is right on Windows too (the hijack is keyed
on the second binder and needs the same user, who can already read the identity
seed — and SO_EXCLUSIVEADDRUSE would block rebinding for 120 s of TIME_WAIT after
Ctrl-C); the pin store never sees CRLF because both halves are binary-mode; and
the truncated SOCKET is type hygiene, not a bug, because Windows handles are
capped near 2^26 and MSDN documents the round trip. Header order, WSAStartup,
BCryptGenRandom, the wire format's byte-for-byte identity and the CMake link line
were all checked and are clean.

Also here:

  - CI, which is what turns "never built with MSVC, never executed" into a
    measured result. The Windows job compiles with MSVC, runs the suites
    natively, and then requires a complete USB Mass Storage exchange over a real
    loopback socket. The connect loop is a retry rather than a port probe on
    purpose: a bare TCP connect is ACCEPTED by the exporter and stalls its
    handshake loop, so probing the port perturbs what it measures. Verified
    locally: attempt 1 exits 3 and pairs, attempt 2 exits 0 with RESULT=PASS.
  - test_macipc split at #if !defined(_WIN32). It mixed a portable codec suite
    with a POSIX socket suite, so `cmake --build` could not have completed on
    Windows at all. The codec keeps its Windows coverage instead of the suite
    being dropped.
  - main() takes char* argv[], the form the standard actually names.
  - The Windows target now compiles warning-free under the flags CMake configures
    (-Wpedantic -Wconversion -Wsign-conversion -Wshadow), not merely under the
    -Wall -Wextra the cross-build script passes. Those were different claims and
    only the weaker one was true.

+++ airusb/core/Platform.h
@@ -63,7 +63,11 @@ inline bool connectionLost(int e) noexcept
 inline bool setNonBlocking(SocketHandle s) noexcept
 {
     u_long on = 1;
-    return ::ioctlsocket(s, FIONBIO, &on) == 0;
+    // FIONBIO is defined as an unsigned long (0x8004667E) but ioctlsocket takes a
+    // signed long, so the canonical MSDN call is a sign-changing conversion. The
+    // bit pattern is what the API matches on and it is preserved; the cast just
+    // stops -Wsign-conversion reporting the value change as news.
+    return ::ioctlsocket(s, static_cast<long>(FIONBIO), &on) == 0;
 }
 
 /// Winsock needs initialising once per process, and is refcounted, so calling
